feat: Allow specifying a path to a custom js function (#267)
This PR allows the user to write their custom js function in a separate
file instead of inside the config.yaml. This cleans up the config when
having long custom functions as well as making it possible for the user
to have full IDE support when writing the actual javascript.