fix: `<a target="_blank">` security vulnerability (#413)
Apparently, in old browsers, using just `target="_blank"` to open a web
page in a new tab gives access to some parts of the originating tab. See
this link for an up to date account of the problem:
https://mathiasbynens.github.io/rel-noopener/

TL;DR: [Just also add ` rel="noopener noreferrer"` to be on the safe
side](https://stackoverflow.com/a/17711167/2352071)

Or if you actually need exactly that behaviour (in newer browsers that
changed the default), use `rel="opener"`. With regard to this: I am not
sure about the change in the template file. Maybe that behavior is
somehow wanted there?