chore: Add javascript functions for static webview (#482)
This PR updates the datavzrd javascript library to export some functions
that are needed for the static webview. This makes the static webview
and datavzrd itself rely on the exact same js therefore heavily reducing
any maintenance work needed. When this is merged we can start
automatically updating the bundle.js in the view repository for every
datavzrd release.