Support the `ellipsis` setting on pill plots.
The logic for `renderPill` is similar to the existing logic in
`shortenColumn`. However, the pills require different wrapping elements,
so it was easier to duplicate some of this logic than to try and unify
these utilities.

The new logic also supports pills with `ellipsis === 0`, in which case
the pills are rendered as circles without any text content, and the pill
value is shown in a tooltip on hover.