feat: Add --entry option for datavzrd publish command (#879)
This pull request introduces a new feature to specify an entry point for
the report in the `publish` command. The changes primarily involve
adding an optional `entry` parameter to various structs and methods, and
implementing logic to handle this new parameter. Fixes #874

### New Feature: Entry Point for Report

*
[`src/cli.rs`](diffhunk://#diff-b2812f19576dd53d0c35b107a322f58a00fce6977f4b5976e1961853982af3ccR61-R64):
Added an optional `entry` parameter to the `publish` command in the
`Command` enum.
*
[`src/main.rs`](diffhunk://#diff-42cb6807ad74b3e201c5a7ca98b911c5fa08380e942be6e4ac5807f8377f87fcR37-R39):
Updated the `publish` command handling to pass the new `entry` parameter
to the `Repository` struct.
* `src/publish.rs`: 
* Added an `entry` field to the `Repository` struct and updated the
`new` method to accept this parameter.
* Implemented the `entry` method to handle the entry point logic,
creating a redirect file if the entry is different from `index.html`.
* Called the `entry` method during the repository publishing process to
ensure the entry point is handled correctly.